% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{constructGseaResultTable}
\alias{constructGseaResultTable}
\title{Builds the table of GSEA statistics to present to the user.}
\usage{
constructGseaResultTable(
  mg,
  method = sparrow::resultNames(mg)[1L],
  fdr = 0.1,
  prioritize = NULL
)
}
\arguments{
\item{mg}{`SparrowResult` object}

\item{method}{the method to show statistics for}

\item{fdr}{the FDR cut off to present statistics for}

\item{prioritize}{the preferred collections to put at the top of the
list. The collection column of the table is turned into a factor for
more useful display with datatable's filter. Specifcying collections
here will put those collections at the front of the factor levels and
therofre prioritize their display in the select dropdown for the filter}
}
\value{
a data.table of the statistics that match the filtering criteria.
  A 0-row data.table is returned if nothing passes.
}
\description{
Creates a table gene set statistics genereated from a GSEA `method` for all
gene sets with an FDR (by collection) less than given threshold.
}
\examples{
sres <- sparrow::exampleSparrowResult()
constructGseaResultTable(sres)
}
