% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{exampleUISetup}
\alias{exampleUISetup}
\title{Generates the CSS/UI setup to reduce code required in documentation examples}
\usage{
exampleUISetup()
}
\value{
inlined html/css
}
\description{
Includes some boilerplate code required in the shiny app UI setup that we
don't want to repeat over and over.
}
\details{
For internal function only, but exported so we can use it in examples.
}
