% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSetContrastView-module.R
\name{geneSetContrastViewUI}
\alias{geneSetContrastViewUI}
\alias{geneSetContrastView}
\alias{is.geneSetContrastViewer}
\alias{updateActiveGeneSetInContrastView}
\title{A module to encapsulate browsing differential statistics of a geneset.}
\usage{
geneSetContrastViewUI(id, height = "590px", width = "400px")

geneSetContrastView(
  input,
  output,
  session,
  mgc,
  server = TRUE,
  maxOptions = Inf,
  sep = "_::_",
  feature.link.fn = ncbi.entrez.link,
  feature_table_filter = "none",
  itools = c("wheel_zoom", "box_select", "reset", "save")
)

is.geneSetContrastViewer(x)

updateActiveGeneSetInContrastView(session, viewer, geneset, mgc)
}
\arguments{
\item{id}{the shiny id of the module}

\item{height, width}{the height and width of the module}

\item{input}{the shiny-required bits for the module}

\item{output}{the shiny-required bits for the module}

\item{session}{the shiny-required bits for the module}

\item{mgc}{A [SparrowResultContainer()] object}

\item{server}{boolean to indicate whether the genesets in the geneSetSelect
widget should be rendered server side or not (Default: `TRUE`)}

\item{maxOptions}{a paremeter used to customize the
`GeneSetSelect::selectizeInput` UI element to configure the maximum number
of elements to include in the select dropdown, the remainder of the
genesets will be loaded from the server side. Default: `Inf` for all.}

\item{sep}{the string used to concatenate geneset `collection` and `name`
to generate a uniqe string for a geneset}

\item{viewer}{a `geneSetContrastViewModule`}

\item{geneset}{the key of a geneset to select}
}
\value{
A `tagList` of html stuff to dump into the UI.

the `geneSetContrastView` module returns a reactive list,
  with a `$gs` element that indicates the currently active geneset in
  the `geneSetSelect` module, and a \code{$selected} element, a character
  vector of feature_ids currently brushed in a contrast view.
}
\description{
The module is meant to be displayed in "a box" so that the user can examine
the coherent (or not) behavior of the geneset across a contrast with respect
to the background distribution of all genes in the contrast.
}
\details{
Embedded within this module is the \code{\link{geneSetSelect}} module, which
provides the list of genesets the user can examine, as well as the title of
the current geneset under scrutiny.

Below the geneSet picker, we embed an \code{\link{iplot}} so that the
user can observe the behavior of the geneset across the contrast. The user
can pick the type of plot to show (density or boxplot) as well as which
statistics to use for plotting (logFC or t-statistics).

A `updateActiveGeneSetInContrastView` function is provided to enable
interactions external to this module the ability to update the geneset
selected in the [geneSetSelect()] module.
}
\section{Functions}{
\itemize{
\item \code{geneSetContrastView}: creates a server module that includes
a geneset select dropdown, and a box to view the graphical result of the
GSEA contrast

\item \code{is.geneSetContrastViewer}: Internal function that tests if
an object has all the shiny bits to be a geneSetContrastView.

\item \code{updateActiveGeneSetInContrastView}: allows you to update the active
geneset in a contrast view module externally.
}}

