% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactiveGeneSetDb-module.R
\name{GeneSetDb.ReactiveGeneSetDb}
\alias{GeneSetDb.ReactiveGeneSetDb}
\title{Returns the filtered unreactive GeneSetDb from the reactiveGeneSetDb module.}
\usage{
GeneSetDb.ReactiveGeneSetDb(x, ...)
}
\arguments{
\item{x}{A ReactiveGeneSetDb}

\item{...}{pass through (not used)}
}
\value{
A standard / static GeneSetDb object
}
\description{
The user can choose to include subsets of collections, as well as
genesets of a certain size.
}
\details{
This can only be run within a reactive context.
}
