% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSetSelect-module.R
\name{geneSetSelectUI}
\alias{geneSetSelectUI}
\alias{geneSetSelect}
\alias{updateGeneSetSelect}
\alias{gs.render.select.ui}
\alias{gs.select.choices}
\title{A module that creates a dynamic selectizeInput for a SparrowResult object}
\usage{
geneSetSelectUI(id, label = "Select Gene Set")

geneSetSelect(
  input,
  output,
  session,
  mgc,
  server = TRUE,
  maxOptions = Inf,
  sep = "_::_"
)

updateGeneSetSelect(
  session,
  id,
  label = NULL,
  choices = NULL,
  selected = NULL,
  options = list(),
  server = FALSE
)

gs.render.select.ui(
  ns,
  choices,
  server = TRUE,
  maxOptions = 1000,
  sep = "_::_"
)

gs.select.choices(mg, sep = "_::_")
}
\arguments{
\item{id}{the 'naked' module id}

\item{label}{The label to set for the input object.}

\item{input, output, session}{the shiny-required bits for the module}

\item{mgc}{A [SparrowResultContainer()] object}

\item{server}{boolean to indicate whether the genesets in the geneSetSelect
widget should be rendered server side or not (Default: `TRUE`)}

\item{maxOptions}{a paremeter used to customize the
`GeneSetSelect::selectizeInput` UI element to configure the maximum number
of elements to include in the select dropdown, the remainder of the
genesets will be loaded from the server side. Default: `Inf` for all.}

\item{sep}{the string used to concatenate geneset `collection` and `name`
to generate a uniqe string for a geneset}

\item{choices}{the output of `gs.select.choices(SparrowResult)`}

\item{selected}{The initially selected value (or multiple values if \code{multiple = TRUE}). If not specified then defaults to the first value for
single-select lists and no values for multiple select lists.}

\item{options}{A list of options. See the documentation of \pkg{selectize.js}
for possible options (character option values inside \code{\link[base:AsIs]{base::I()}} will
be treated as literal JavaScript code; see \code{\link[shiny:renderDataTable]{renderDataTable()}}
for details).}

\item{ns}{the namespace function for this module}

\item{mg}{`SparrowResult` to build options for}
}
\value{
a reactive list of reactive elements

a properly wired `[shiny::selectizeInput()]` UI element.

`data.table` to populate `choices` of `selectizeInput`
}
\description{
This code was inspired from the
\href{https://gist.github.com/MarkEdmondson1234/7e56ee7ac5caa74224327489b0849e61}{dynamicSelectShinyModule.R}
gist.
}
\section{Functions}{
\itemize{
\item \code{updateGeneSetSelect}: updates an "external" geneSetSelectModule
with new choices

\item \code{gs.render.select.ui}: Internal function to build a `selectizeInput`
widget that is specific to a SparrowResult.

\item \code{gs.select.choices}: Internal function to build a `data.frame`
  used to populate geneset choices for a select input.

Note that when returning a data.frame for the choices from
`gs.select.choice()`, we need a column called `"value"` and a column called
`"label"`.

* `value`: the value that is sent back when an item is selected
* `label`: athe text that appears in the selection after its triggered
}}

\section{Module Return}{

Returns information about the `geneSetSelect` object
}

