% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-facile-internal.R
\name{initialized}
\alias{initialized}
\title{Tests if a module is initialized}
\usage{
initialized(x, ...)
}
\arguments{
\item{x}{A thing used in a shiny environment that requires time to initialize
before it can be used.}

\item{...}{pass through}
}
\value{
logical TRUE/FALSE
}
\description{
Tests if a module is initialized
}
