% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgTableBrowser-module.R
\name{mgTableBrowser}
\alias{mgTableBrowser}
\alias{mgTableBrowserUI}
\title{A module that displays an interactive table of genesets from an analysis.}
\usage{
mgTableBrowser(input, output, session, mgc, method, fdr, server = TRUE)

mgTableBrowserUI(id)
}
\arguments{
\item{input, output, session}{shiny module bits}

\item{mgc}{the \code{SparrowResultContainer} object}

\item{method}{a reactive that determines which method to explore from this
result}

\item{fdr}{a reactive that gives the fdr threshold to filter results in the
table by.}

\item{server}{boolean passed to [DT::renderDataTable()] (default: `TRUE`).}

\item{id}{the shiny id of the UI module}
}
\value{
a list with reactives:
\describe{
  \item{$stats}{
    The table of gene sets and their statistics that pass the prescribed
    \code{fdr} thershold
  }
  \item{$selected}{
    The geneset that is currently "active"/selected by the user. This
    is defined as \code{<collection>_::_<name>}
  }
}
}
\description{
This module lists all of the genesets from an analysis that are significant
given an `fdr` and `method` of analysis. Row selections trigger a shiny
event that broadcasts the key of the geneset that was selected. You would
then want to `observeEvent(this$selected)` in your `server.R` (or similar)
so you can react to the gene set selection the user triggers.
}
\section{Functions}{
\itemize{
\item \code{mgTableBrowserUI}: The UI for the module.
}}

