% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgVolcano-module.R
\name{mgVolcano}
\alias{mgVolcano}
\alias{mgVolcanoUI}
\title{A module to visualize a volcano plot.}
\usage{
mgVolcano(
  input,
  output,
  session,
  x,
  stats = "dge",
  xaxis = "logFC",
  yaxis = "pval",
  idx = "idx",
  tools = c("box_select", "reset", "save"),
  width = NULL,
  height = NULL,
  highlight = reactive(NULL),
  default_xhex = 1,
  default_yhex = 0.1,
  webgl = FALSE,
  ...
)

mgVolcanoUI(
  id,
  x,
  stats = "dge",
  xaxis = "logFC",
  yaxis = "padj",
  idx = "idx",
  hexbin = TRUE,
  default_xhex = 1,
  default_yhex = 0.1
)
}
\arguments{
\item{input, output, session}{shiny bits}

\item{x}{A \code{SparrowResult} object, or a \code{data.frame}}

\item{stats}{One of \code{"dge"} or \code{resultNames(x)}}

\item{xaxis}{the column of the the provided (or extracted)
\code{data.frame} to use for the xaxis and yaxis of the volcano}

\item{yaxis}{the column of the the provided (or extracted)
\code{data.frame} to use for the xaxis and yaxis of the volcano}

\item{idx}{The column of the \code{data.frame} to use as the identifier
for the element in the row. You probably don't want to mess with this}

\item{tools}{the plotly tools to include in the figures toolbar}

\item{width}{the width and height of the output plotly plot}

\item{height}{the width and height of the output plotly plot}

\item{highlight}{a reactive vector of featureIds that indicate which points
to highlight on the volcano, irresespective of their "hexbin" status.}

\item{default_xhex, default_yhex}{default values to set the slider inputs
to pass in for hex-ization}

\item{webgl}{use webgl to make the plotly plot}

\item{...}{pass through arguments (not used)}

\item{id}{the shiny id of the output widget}
}
\value{
A reactive data.frame that includes the information of the genes
  that are brushed in the volcano plot. If no genes have been selected, then
  `NULL`.
}
\description{
Users can specify logFC and FDR thresholds for which the individual dots
in the volcano are rather hexbins so that the browser doesn't have to handle
a single point for each gene/dot.
}
\section{Functions}{
\itemize{
\item \code{mgVolcanoUI}: the UI for the module
}}

