% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{renderFeatureStatsDataTable}
\alias{renderFeatureStatsDataTable}
\title{Create a DT::datatable of feature level statistics for use in shiny bits}
\usage{
renderFeatureStatsDataTable(
  x,
  features = NULL,
  digits = 3,
  columns = NULL,
  feature.link.fn = NULL,
  order.by = "logFC",
  order.dir = c("desc", "asc"),
  filter = "none",
  length.opts = c(10, 25, 50, 100, 250)
)
}
\arguments{
\item{x}{A `SparrowResult` or `data.frame` of feature level
statistics. When `x` is a `SparrowResult`, the
`logFC` feature level statistics will be extracted for
display.}

\item{features}{A character vector that specifies the subset of
`feature_id`'s to display from `x`. If `NULL` (default),
all of `x` will be used.}

\item{digits}{number of digits to round the numeric columns to}

\item{columns}{the columns from `x` to use. If `missing`, then
only `c('symbol', 'feature_id', 'logFC', 'pval', 'padj', order.by)`
will be used. If explicitly set to `NULL` all columns will be used.}

\item{feature.link.fn}{A function that receives the data.frame of statistics
to be rendered and transforms one of its columns into a hyperlink for
further reference. Refer to [ncbi.entrez.link()] as an example.}

\item{order.by}{name of the column to order the geneset stats by}

\item{order.dir}{do we order the table by the values in `order.by` in
`"desc"`-ending, or `"asc"`-ending order (default: `"desc"`)}

\item{filter}{passed to [DT::datatable()]}

\item{length.opts}{parameter passed as an option to [DT::datatable()] that
specifies how long the table can be per page.}
}
\value{
A [DT::datatable()] of feature statistics
}
\description{
We often want to display an interactive table of feature level statistics
for all features. This function is a convenience wrapper to do that.
}
\examples{
sr <- sparrow::exampleSparrowResult()
lfc <- sparrow::logFC(sr)
renderFeatureStatsDataTable(sr, sample(lfc$feature_id, 200),
                            feature.link.fn = ncbi.entrez.link)
}
