% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{renderGseaResultTableDataTable}
\alias{renderGseaResultTableDataTable}
\title{Creates a DT::datatable of geneset level GSEA results for use in shiny bits}
\usage{
renderGseaResultTableDataTable(x, method, mg, digits = 3)
}
\arguments{
\item{x}{The set of GSEA statistics generated from from
[constructGseaResultTable()]}

\item{method}{the GSEA method being used fo rdisplay}

\item{mg}{The `SparrowResult` object. This is used swap in the
URL links for genesets using [sparrow::geneSetURL()].}
}
\value{
a DT::DataTable
}
\description{
Creates a DT::datatable of geneset level GSEA results for use in shiny bits
}
