% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{roundDT}
\alias{roundDT}
\title{Rounds all of the numeric columns of a DT}
\usage{
roundDT(x, digits = 3)
}
\arguments{
\item{x}{a DT::datatable}

\item{digits}{the number of digits to round. If \code{NA}, then no rounding
is performed}
}
\value{
a rounded DT::datatable
}
\description{
This is a convenience function around [DT::formatRound()] that identifies
all of the numeric columns and rounds them, as opposed to just rounding
prespecified columns.
}
\examples{
df <- data.frame(a=rnorm(10), b=sample(letters, 10), c=rnorm(10))
roundDT(DT::datatable(df),  digits=2)
}
