% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparrowResultContainer.R
\name{SparrowResultContainer}
\alias{SparrowResultContainer}
\title{A wrapper for a SparrowResult that promotes internals of object.}
\usage{
SparrowResultContainer(x)
}
\arguments{
\item{x}{A `[sparrow::SparrowResult()]` object, or a path to
an rds-serliazed one}
}
\value{
a `SparrowResultContainer` object (list), with these elements:
\describe{
  \item{sr}{The `SparrowResult` object}
  \item{methods}{A character vector of method names that were run}
}
}
\description{
This creates a facade over `SparrowResult` object that enables us to wrap
`reactive`s around its internal intresting bits that are used elsewhere.
}
\examples{
sres <- sparrow::exampleSparrowResult()
src <- SparrowResultContainer(sres)
}
