% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Interactively explore a SparrowResult via a shiny app}
\usage{
explore(x)
}
\arguments{
\item{x}{A `SparrowResult` object, or path to one as an *.rds. If
missing, the shiny app will load without a `SparrowResult` object
to explore, and the user can upload one into the app.}
}
\description{
This will launch a shiny application that enables exploratory analysis of
the results from the different GSEA methods run within a `SparrowResul`.
}
\details{
Reference the "shiny-sparrow" vignette for more detailed documentation of
the functionality provided by this application.
}
\examples{
\donttest{
vm <- sparrow::exampleExpressionSet()
gdb <- sparrow::exampleGeneSetDb()
sr <- sparrow::seas(vm, gdb, vm$design, methods=c('camera', 'fry'))
explore(sr)
}
}
