% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgGeneSetSummaryByGenes-module.R
\name{mgGeneSetSummaryByGene}
\alias{mgGeneSetSummaryByGene}
\alias{mgGeneSetSummaryByGeneUI}
\title{A module to show a tabular view of genesets that contain genes of interest.}
\usage{
mgGeneSetSummaryByGene(input, output, session, mgc, features, method, fdr)

mgGeneSetSummaryByGeneUI(id, mg = NULL)
}
\description{
Users can look up which gene sets include query genes of interest. This is
useful when you are exploring geneset hits and find esoteric results there.
You can select some of the "more extreme" genes in that geneset, for instance
to see if they belong to another one that makes more sense to you given the
biological context of your experiment.
}
