% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgResultFilter-module.R
\name{mgResultFilter}
\alias{mgResultFilter}
\alias{mgResultFilterUI}
\title{A module that allows users to select which GSEA results to display.}
\usage{
mgResultFilter(input, output, session, mgc)

mgResultFilterUI(id)
}
\arguments{
\item{input, output, }{session shiny bits}

\item{mgc}{the `SparrowResultContainer`}
}
\value{
A list that includes the following reactives:
\describe{
  \item{$method}{The name of the GSEA method selected by the user}
  \item{$fdr}{The fdr threshold specified by the user}
}
}
\description{
The app is setup to only allow users to explore GSEA results one method at a
time. Given a SparrowResultContainer `mgc`, this module presents the user
with a dropdown list of GSEA methods that were run that they want to explore
and an FDR cutoff that limits the gene sets returned for exploration.
}
\section{Functions}{
\itemize{
\item \code{mgResultFilterUI}: the UI for the module: presents a selectInput
for the user to pick which GSEA method to expl
}}

