% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactiveGeneSetSelect-module.R
\name{reactiveGeneSetSelect}
\alias{reactiveGeneSetSelect}
\alias{reactiveGeneSetSelectUI}
\title{A gene set selector that enables users to pre-filter universe to choose from}
\usage{
reactiveGeneSetSelect(input, output, session, gdb = NULL, ...)

reactiveGeneSetSelectUI(id, label = NULL, dropdown_width = "350px", ...)
}
\arguments{
\item{input, output, session}{shiny module bits}

\item{gdb}{A [reactiveGeneSetDb()] module, or `NULL`}

\item{...}{pass through (not used)}

\item{id}{the id/namespace for the module}

\item{label}{passed to [shiny::selectizeInput()]}

\item{dropdown_width}{the width of the dropdown element in pixels}
}
\value{
A list of geneset info and membersihp
}
\description{
This provides the user to pick a single gene set from a GeneSetDb collection.
To help the user find the gene set, the user can select which collections
and gene set sizes that serve as the universe of potential gene sets to
pick from in a GeneSetDb.
}
\section{Functions}{
\itemize{
\item \code{reactiveGeneSetSelectUI}: The UI function
}}

\examples{
sres <- sparrow::exampleSparrowResult()
gdb <- sparrow::geneSetDb(sres)
app <- shiny::shinyApp(
  ui = shiny::shinyUI(shiny::fluidPage(
    exampleUISetup(),
    title = "Gene set selctor",
    reactiveGeneSetSelectUI("mod"))),
  server = function(input, output, session) {
    shiny::callModule(reactiveGeneSetSelect, "mod", gdb)
  })
if (interactive()) {
  shiny::runApp(app)
}
}
