% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{renderGseaResultTableDataTable}
\alias{renderGseaResultTableDataTable}
\title{Creates a DT::datatable of geneset level GSEA results for use in shiny bits.}
\usage{
renderGseaResultTableDataTable(x, method, mg, digits = 3)
}
\arguments{
\item{x}{The set of GSEA statistics generated from from
[constructGseaResultTable()]}

\item{method}{the GSEA method being used fo rdisplay}

\item{mg}{The `SparrowResult` object. This is used swap in the
URL links for genesets using [sparrow::geneSetURL()].}

\item{digits}{the number of digits to round numer columns to.}
}
\value{
A [DT::datatable()] for display
}
\description{
This function is closely tied to the exact output of the GSEA methods from
the [sparrow::seas()].
}
\examples{
sres <- sparrow::exampleSparrowResult()
method <- sparrow::resultNames(sres)[1L]
stable <- constructGseaResultTable(sres, method)
renderGseaResultTableDataTable(stable, method, sres, digits = 2)
}
