% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-SparrowResultContainer.R
\name{SparrowResultContainer}
\alias{SparrowResultContainer}
\title{A wrapper for a SparrowResult that makes some shiny bits explicit}
\usage{
SparrowResultContainer(x)
}
\arguments{
\item{x}{A `[sparrow::SparrowResult()]` object, or a path to
an rds-serliazed one}
}
\value{
a `SparrowResultContainer` object
}
\description{
A wrapper for a SparrowResult that makes some shiny bits explicit
}
