% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userDefinedGeneSetDb-module.R
\name{userDefinedGeneSetDb}
\alias{userDefinedGeneSetDb}
\alias{userDefinedGeneSetDbUI}
\title{Creates a GeneSetDb from a user-specfied gene set definition table.}
\usage{
userDefinedGeneSetDb(input, output, session, ...)

userDefinedGeneSetDbUI(id, ...)
}
\arguments{
\item{input, output, session}{shiny module bits}

\item{...}{pass through arguments}

\item{id}{the module namespace}
}
\value{
A list of reactive components. `$gdb()` will be a GeneSetDb when
  the user uploades a valid gene set definition file. Otherwise it will be
  `NULL`.
}
\description{
This module provides an upload button that allows a user to upload a
table of gene set definitions in [sparrow::GeneSetDb()] format as a CSV file
or excel file (if the readxl package is installed). Minimal validation checks
are implemented.
}
\section{Functions}{
\itemize{
\item \code{userDefinedGeneSetDbUI}: ui for the module
}}

\examples{
# You can upload the file available here:
(ex.fn <- system.file("testdata", "user-defined-genesets-example.xlsx",
                      package = "sparrow.shiny"))
app <- shiny::shinyApp(
  ui = shiny::shinyUI(shiny::fluidPage(
    exampleUISetup(),
    title = "Costum Gene Set Collection Upload",
    userDefinedGeneSetDbUI("mod"))),
  server = function(input, output, session) {
    shiny::callModule(userDefinedGeneSetDb, "mod", gdb)
  })
if (interactive()) {
  shiny::runApp(app)
}
}
